+++ PdfCreator/AbstractPdfCreator.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2021 Heimrich & Hannot GmbH
+ * Copyright (c) 2022 Heimrich & Hannot GmbH
  *
  * @license LGPL-3.0-or-later
  */
@@ -128,11 +128,9 @@ abstract class AbstractPdfCreator
     /**
      * Render the pdf file.
      *
-     * @return string|void
-     *
      * @throw MissingDependenciesException
      */
-    abstract public function render();
+    abstract public function render(): PdfCreatorResult;
 
     /**
      * @return string